refactor: (#281) WebRequest → Request 변경
